Queue input events at the queued cursor, not the mouse state
QueueInputEventAtCursor read the raw mouse cursor record where the
original reads the queued input position the event manager maintains, so
synthetic events carried a position that the rest of the input path had
not agreed on.  The event queue's coordinate, type and value parameters
are ints, which is also why the polled translator hands the joystick
sample straight through instead of truncating it to a word.

QueueInputEvent ORs the shift modifier in rather than assigning it, so a
shift held together with control or a mouse button no longer discarded
the other bits.

+++ include/wc1funcs.h
@@ -1534,14 +1534,13 @@ void ShutdownEventManager(void);                                      /* 0x4625C
 unsigned short __stdcall ConfigureEventManagerPointer(
     unsigned char *shape, short frame);                               /* WC2 unmapped */
 void __stdcall SetEventManagerPump(void (*pump)(void));                /* WC2 unmapped */
-void TranslatePolledInputEvent(unsigned short type,
+void TranslatePolledInputEvent(unsigned int type,
                                unsigned int value);                   /* 0x462625 */
 void QueueInputEventAtCursor(unsigned int type, short primaryButton,
                              short secondaryButton);                   /* 0x462849 */
 InputEvent *AllocateInputEvent(void);                                  /* 0x462890 */
 void ReleaseInputEvent(InputEvent *event);                             /* 0x46294F */
-void QueueInputEvent(unsigned short type, unsigned short x,
-                     unsigned short y, unsigned short value,
+void QueueInputEvent(int type, int x, int y, int value,
                      int primaryButton, int secondaryButton,
                      unsigned int ignored, unsigned int field14,
                      unsigned int field18);                            /* 0x4629A7 */
